Fix Right-Click Context Menu (#1647)
Added the disable and cache options to the right-click context menu. Working on fixing the bugs that prevent the navigation component and context menu operator status to sync and making sure grouped and ungrouped operators reflect the accurate disable/enable or cache/remove cache options in the menu.
