+++ core/new-gui/src/app/workspace/component/workflow-editor/context-menu/context-menu/context-menu.component.scss
@@ -1,0 +1,11 @@
+li {
+  padding-top: 2px;
+  padding-bottom: 2px;
+  font-size: 14px;
+  min-width: 120px;
+
+  span {
+    margin-left: 2px;
+    margin-right: 12px;
+  }
+}
