+++ core/new-gui/src/app/workspace/component/workflow-editor/workflow-editor.component.html
@@ -1,8 +1,13 @@
 <div style="position: relative; width: 100%; height: 100%">
   <div
     [attr.id]="WORKFLOW_EDITOR_JOINTJS_WRAPPER_ID"
+    (contextmenu)="contextMenu($event, menu)"
     class="texera-workspace-workflow-editor-body texera-workflow-component-body"
   >
     <div [attr.id]="WORKFLOW_EDITOR_JOINTJS_ID" class="texera-workspace-workflow-jointjs"></div>
   </div>
 </div>
+
+<nz-dropdown-menu nzNoAnimation #menu="nzDropdownMenu">
+  <context-menu></context-menu>
+</nz-dropdown-menu>
