+++ core/new-gui/src/app/workspace/component/workflow-editor/workflow-editor.component.spec.ts
@@ -35,6 +35,7 @@ import { StubUserService } from "src/app/common/service/user/stub-user.service";
 import { WorkflowVersionService } from "src/app/dashboard/service/workflow-version/workflow-version.service";
 import { WorkflowCollabService } from "../../service/workflow-collab/workflow-collab.service";
 import { of } from "rxjs";
+import { NzContextMenuService, NzDropDownModule } from "ng-zorro-antd/dropdown";
 
 describe("WorkflowEditorComponent", () => {
   /**
@@ -51,7 +52,7 @@ describe("WorkflowEditorComponent", () => {
       waitForAsync(() => {
         TestBed.configureTestingModule({
           declarations: [WorkflowEditorComponent],
-          imports: [HttpClientTestingModule, NzModalModule],
+          imports: [HttpClientTestingModule, NzModalModule, NzDropDownModule],
           providers: [
             JointUIService,
             WorkflowUtilService,
@@ -60,6 +61,7 @@ describe("WorkflowEditorComponent", () => {
             ResultPanelToggleService,
             ValidationWorkflowService,
             WorkflowActionService,
+            NzContextMenuService,
             Overlay,
             {
               provide: OperatorMetadataService,
@@ -152,7 +154,7 @@ describe("WorkflowEditorComponent", () => {
       waitForAsync(() => {
         TestBed.configureTestingModule({
           declarations: [WorkflowEditorComponent, NzModalCommentBoxComponent],
-          imports: [HttpClientTestingModule, NzModalModule, NoopAnimationsModule],
+          imports: [HttpClientTestingModule, NzModalModule, NzDropDownModule, NoopAnimationsModule],
           providers: [
             JointUIService,
             WorkflowUtilService,
@@ -162,6 +164,7 @@ describe("WorkflowEditorComponent", () => {
             ValidationWorkflowService,
             DragDropService,
             NzModalService,
+            NzContextMenuService,
             {
               provide: OperatorMetadataService,
               useClass: StubOperatorMetadataService,
